Cast response body to string so a non-zero length string is returned
getContents gets the contents of the body from the current position in
the response stream to the end. The call to formatResponse (and
subsequent casting to string in that function) moves the position in the
stream to the end.

The behaviour of __toString on StreamInterface dictates seeking to the
beginning of the stream before reading, and this solves the problem.